requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
appnope==0.1.3
argcomplete==3.1.6
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
blinker==1.7.0
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.4
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.2.0
Covid19AnalysisProjectLib @ file:///Users/alexis/Cranfield/ml-big-data/Assessment
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
executing==2.0.0
findspark==2.0.1
fiona==1.9.5
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.44.0
frozenlist==1.4.0
geopandas==0.14.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prison==0.2.1
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.1
pyspark==3.5.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
